va-attorney-agent/requirements.txt
anthropic>=0.42.0
python-dotenv>=1.0.0
httpx>=0.27.0
